Use build tags when starting logs agent to reduce dependencies in serverless.
