+++ pkg/logs/logs_not_serverless_test.go
@@ -3,6 +3,9 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
+//go:build !serverless
+// +build !serverless
+
 package logs
 
 import (
@@ -11,15 +14,8 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestBuildServerlessEndpoints(t *testing.T) {
-	endpoints, err := buildEndpoints(true)
-	assert.Nil(t, err)
-	assert.Equal(t, "http-intake.logs.datadoghq.com", endpoints.Main.Host)
-	assert.Equal(t, "lambda-extension", string(endpoints.Main.Origin))
-}
-
 func TestBuildEndpoints(t *testing.T) {
-	endpoints, err := buildEndpoints(false)
+	endpoints, err := buildEndpoints()
 	assert.Nil(t, err)
 	assert.Equal(t, "agent-intake.logs.datadoghq.com", endpoints.Main.Host)
 }
